+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * refactor: rename `createFrom()` to `with()`
 * refactor: rename the term 'safe' to 'large'
 * refactor(`RecollectionDecorator`): rename `withCriteria` to `applyCriteria`
+* refactor: rename the term 'large' to 'basic'
 
 ## 0.3.0